EcoFriendlyRouteProject/requirements.txt
flask
flask-cors
requests
joblib
scikit-learn
pandas
numpy
python-dotenv
